NGUYEN HOANG PHUC

+++ 2.py
@@ -4,15 +4,15 @@ print("Hello,", name)
 
 import math
 rad = input("What is radius of circle?")
-r = int(rad)
+r = float(rad)
 result = math.pi*r**2
 print(f"The area of the circle: , {result:2.2f}")
 
 import math
 length = input("What is the length of the rectangle?")
-l = int(length)
+l = float(length)
 width = input("What is the width of the rectangle?")
-w = int(width)
+w = float(width)
 pe = (l+w)*2
 print("The perimeter of the rectangle is:", pe)
 area = (l*w)
